With magic crutch in algorithm

+++ src/extra/baserin.py
@@ -34,9 +34,8 @@ class BaseRin:
     time_to_reconnect = cfg_data.get('time to reconnect')
     orders_depth = cfg_data.get('orders depth')
 
-    # dtype_float64 = np.float64
-    dtype_float64 = float
-    dtype_int64 = int
+    dtype_float64 = np.float_
+    dtype_int64 = np.int_
 
     work_dir = utils.get_dir('rin-bot')
     blacklist = utils.create_empty_file(
